# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  images: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  sizes?: string[];
  stock?: number;
  inStock: boolean;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt: string;
}

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  comparePrice?: number;
  images: string[];
  categoryId: number;
  sizes?: string[];
  stock?: number;
  badge?: string;
  sku?: string;
}

export interface ProductUpdate {
  title?: string;
  description?: string;
  price?: number;
  /** @nullable */
  comparePrice?: number | null;
  images?: string[];
  categoryId?: number;
  sizes?: string[];
  stock?: number;
  /** @nullable */
  badge?: string | null;
  sku?: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
}

export interface CartItem {
  id: number;
  productId: number;
  productTitle: string;
  productImage: string;
  price: number;
  quantity: number;
  /** @nullable */
  size: string | null;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  subtotal: number;
  total: number;
  itemCount?: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
  size?: string;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productTitle: string;
  /** @nullable */
  productImage?: string | null;
  price: number;
  quantity: number;
  /** @nullable */
  size?: string | null;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  postalCode?: string | null;
  total: number;
  status: string;
  paymentStatus?: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  netopiaOrderId?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  sessionId: string;
  shippingAddress: string;
  city: string;
  county: string;
  postalCode: string;
  notes?: string;
}

export type OrderCreatedResponseNetopiaFormData = {[key: string]: string};

export interface OrderCreatedResponse {
  orderId: number;
  paymentUrl: string;
  netopiaFormData?: OrderCreatedResponseNetopiaFormData;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface WishlistItem {
  id: number;
  productId: number;
  product: Product;
}

export interface WishlistToggle {
  sessionId: string;
  productId: number;
}

export interface WishlistToggleResponse {
  added: boolean;
  productId: number;
}

export interface Review {
  id: number;
  productId: number;
  reviewerName: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  verified?: boolean;
  createdAt: string;
}

export interface ReviewInput {
  reviewerName: string;
  reviewerEmail?: string;
  rating: number;
  comment?: string;
}

export interface NewsletterSubscription {
  email: string;
  name?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface AdminStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  pendingOrders: number;
  todayOrders: number;
  todayRevenue: number;
  lowStockProducts: number;
  recentOrders: Order[];
}

export interface InventoryItem {
  id: number;
  title: string;
  /** @nullable */
  sku: string | null;
  stock: number;
  price: number;
  /** @nullable */
  categoryName: string | null;
  /** @nullable */
  badge?: string | null;
  images?: string[];
}

export type ListProductsParams = {
category?: string;
search?: string;
badge?: string;
minPrice?: number;
maxPrice?: number;
inStock?: boolean;
sortBy?: string;
page?: number;
limit?: number;
};

export type GetCartParams = {
sessionId: string;
};

export type ListOrdersParams = {
status?: string;
page?: number;
limit?: number;
};

export type GetWishlistParams = {
sessionId: string;
};

export type NetopiaCallbackBody = {
  env_key?: string;
  data?: string;
};