# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ank's Boutique API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart contents by session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add an item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Cart item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created with payment redirect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreatedResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Wishlist
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscription"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Admin dashboard
...
  /payments/netopia/callback:
    post:
      operationId: netopiaCallback
      tags: [orders]
      summary: Netopia payment IPN callback
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                env_key:
                  type: string
                data:
                  type: string
      responses:
        "200":
          description: Payment processed
          content:
            application/xml:
              schema:
                type: string
components:
  schemas:
...
    Cart:
      type: object
      required: [sessionId, items, subtotal, total]
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        total:
          type: number
        itemCount:
          type: integer
    CartItem:
      type: object
      required: [id, productId, productTitle, productImage, price, quantity, size]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        productImage:
          type: string
        price:
          type: number
        quantity:
          type: integer
        size:
          type: ["string", "null"]
    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
        size:
          type: string
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
    Order:
      type: object
      required: [id, customerName, customerEmail, customerPhone, total, status, items, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        county:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        total:
          type: number
        status:
          type: string
        paymentStatus:
          type: string
        paymentMethod:
          type: ["string", "null"]
        netopiaOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, productTitle, price, quantity]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        productImage:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        size:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [customerName, customerEmail, customerPhone, sessionId, shippingAddress, city, county, postalCode]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        sessionId:
          type: string
        shippingAddress:
          type: string
        city:
          type: string
        county:
          type: string
        postalCode:
          type: string
        notes:
          type: string
    OrderCreatedResponse:
      type: object
      required: [orderId, paymentUrl]
      properties:
        orderId:
          type: integer
        paymentUrl:
          type: string
        netopiaFormData:
          type: object
          additionalProperties:
            type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    OrderListResponse:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    NewsletterSubscription:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: string
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...