# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  id: number;
  productId: number;
  productTitle: string;
  productImage: string;
  price: number;
  quantity: number;
  /** @nullable */
  size: string | null;
}