# lib/api-zod/src/generated/types/newsletterSubscription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscription {
  email: string;
  name?: string;
}