# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  postalCode?: string | null;
  total: number;
  status: string;
  paymentStatus?: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  netopiaOrderId?: string | null;
  items: OrderItem[];
  createdAt: string;
}