# lib/api-zod/src/generated/types/orderCreatedResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCreatedResponseNetopiaFormData } from './orderCreatedResponseNetopiaFormData';

export interface OrderCreatedResponse {
  orderId: number;
  paymentUrl: string;
  netopiaFormData?: OrderCreatedResponseNetopiaFormData;
}